Create BotClean Large
The task was to find the next move of a robot that is positioned in a certain cell of a matrix. The robot can perform five types of operations: LEFT, RIGHT, UP, DOWN, CLEAN. It should clean the dirty cells. The matrix shows the dirty cells with an 'd' and the clean cells with a '-'. The robot itself is represented with a 'b' in the matrix. I used a helper function isDirty(board, i, j) to check if a certain cell is dirty or not. Another helper function printNextMove(nextx, nexty) is used to print the next move.

The strategy i used to solve this task was to look for a dirty cell around the current cell and if there is one, to print the next move to that cell. If there is no dirty cell around the current cell, the robot should move DOWN.

The time complexity of this approach is O(n^2) and the space complexity is O(1).