Create BotClean Partially Observable
The bot is given a 5 * 5 grid. It can move in 4 directions, i.e., left, right, up, and down. The bot is initially positioned at the top left corner of the grid with coordinates (0, 0). The bot is given a partially observable environment. The sensors visibility is confined to its 8 adjacent cells. The bot's goal is to clean all the dirty cells as quickly as possible. The bot's score is (200 - #bot moves)/25.

The bot starts moving in the direction of the nearest dirty cell. If there is a dirty cell in the bot's current position, it cleans the cell. If there are no dirty cells in the bot's current position, it moves in the direction of the nearest dirty cell.